requirements.txt
biopython>=1.72
pandas>=0.23.0
scipy>=1.1.0
seaborn>=0.9.0